Number Input fixes
